renamed components to src dir
