+++ pages/consult.tsx
@@ -1,12 +1,12 @@
 import React from "react";
-import Layout from "../components/layout";
+import Layout from "../src/layout";
 
 const Booking = () => {
-    return (
-        <Layout>
-            <h2>Book a Consultation Appointment</h2>
-        </Layout >
-    );
+  return (
+    <Layout>
+      <h2>Book a Consultation Appointment</h2>
+    </Layout>
+  );
 };
 
 export default Booking;
