+++ pages/index.tsx
@@ -7,11 +7,11 @@ import "swiper/css/pagination";
 import "swiper/css/free-mode";
 
 import { AppImage } from "../types/image";
-import Modal from "../components/modal";
-import Hero from "../components/hero";
-import Layout from "../components/layout";
-import Listing from "../components/listing";
-import CustomImage from "../components/custom-image";
+import Modal from "../src/modal";
+import Hero from "../src/hero";
+import Layout from "../src/layout";
+import Listing from "../src/listing";
+import CustomImage from "../src/custom-image";
 import { extractImages } from "../helpers";
 import { featuredImages } from "../data/featured-images";
 import { serviceTypes } from "../data";
