+++ pages/services/[service-type]/[service-id]/index.tsx
@@ -6,17 +6,17 @@ import {
   SelectExtensionLength,
   SelectHairLength,
   SelectSchedule,
-} from "../../../../components/booking/index";
-import Layout from "../../../../components/layout";
+} from "../../../../src/booking/index";
+import Layout from "../../../../src/layout";
 import { Service } from "../../../../types/service";
-import Button from "../../../../components/button";
+import Button from "../../../../src/button";
 import {
   braidsAndTwistsServices,
   dreadlocksServices,
   strings,
 } from "../../../../data";
 import { bookingMachine } from "../../../../context/booking-machine";
-import { useBookingService } from "../../../../components/hooks/useBookingService";
+import { useBookingService } from "../../../../src/hooks/useBookingService";
 
 interface Props {
   service: Service;
