+++ pages/services/[service-type]/index.tsx
@@ -1,7 +1,7 @@
 import { useRouter } from "next/router";
 import React from "react";
-import Layout from "../../../components/layout";
-import Listing from "../../../components/listing";
+import Layout from "../../../src/layout";
+import Listing from "../../../src/listing";
 import {
   braidsAndTwistsServices,
   dreadlocksServices,
