+++ src/booking/confirm-exit.tsx
