+++ src/booking/confirmation.tsx
