+++ src/custom-image.module.scss
