+++ src/custom-image.tsx
