+++ src/hero.tsx
@@ -1,6 +1,6 @@
 import React from "react";
 import Image from "next/image";
-import Button from "../components/button";
+import Button from "./button";
 import styles from "./hero.module.scss";
 import { AppImage } from "../types/image";
 import { useRouter } from "next/router";
