+++ src/logo/loctineer-desktop-logo.js
