+++ src/logo/loctineer-mobile-logo.js
